# Extract common cache data preparation logic to new method