Refactor the `multiSet` and `multiAdd` methods by extracting the common logic for preparing cache data into a new private method called `prepareCacheData`.