@samdark 2023-10-19
That's a good refactor but it's a cache and one more function call may matter on many-many cache calls. Are you willing to compare performance?