@salehhashemi1992 2023-10-19
I performed a performance test to check the impact of the new method. I ran the test in a loop with 1 million iterations:
```
        $start = microtime(true);

        for ($i = 0; $i < 1000000; $i++) {
            $cache->flush();

            $cache->multiSet([
                'string_test' => 'string_test',
                'number_test' => 42,
                'array_test' => ['array_test' => 'array_test'],
            ], 0);
        }

        $end = microtime(true);

        echo "Old method elapsed time: " . ($end - $start) . " seconds\n";
```
I ran the tests three times:

- ...New method elapsed time: 8.5871288776398 seconds
- ...New method elapsed time: 8.7551159858704 seconds
- ...New method elapsed time: 8.8505420684814 seconds
- 
- ...Old method elapsed time: 8.3703300952911 seconds
- ...Old method elapsed time: 8.3308370113373 seconds
- ...Old method elapsed time: 8.4228310585022 seconds

I think the refactored method is slightly slower, but the difference is minimal and likely won't significantly impact real-world performance.

These tests were done using an array cache. In other cache configs, the time taken for network or disk I/O operations would be much greater than in-memory array cache. Therefore, the overhead of an additional function call (as in your refactor) is likely negligible compared to the time spent on these I/O operations.

@samdark